Fix rubocop warnings
